requirements.txt
cycler==0.10.0
kiwisolver==1.2.0
matplotlib==3.3.0
numpy==1.19.1
opencv-python==4.3.0.36
Pillow==7.2.0
pycryptodome==3.9.8
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
ttkthemes==3.1.0
